+++ lib/src/core/easy_line.dart
@@ -22,6 +22,9 @@ class EasyLine extends StatelessWidget {
   /// Line Axis.
   final Axis axis;
 
+  /// Only for [LineType.normal]. Default: `BorderRadius.circular(100)`
+  final BorderRadius? borderRadius;
+
   const EasyLine({
     key,
     this.length = 50.0,
@@ -31,6 +34,7 @@ class EasyLine extends StatelessWidget {
     this.width = 2.0,
     this.lineType = LineType.dotted,
     this.axis = Axis.horizontal,
+    this.borderRadius,
   }) : super(key: key);
 
   @override
@@ -54,7 +58,7 @@ class EasyLine extends StatelessWidget {
       decoration: lineType == LineType.normal
           ? BoxDecoration(
               color: color,
-              // borderRadius: BorderRadius.circular(100),
+              borderRadius: borderRadius ?? BorderRadius.circular(100),
             )
           : null,
       child: lineType == LineType.dotted